--main/--main/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
PyQt5>=5.15.0
opencv-python>=4.7.0
transformers>=4.28.0
ultralytics>=8.0.0  # 如果使用 YOLOv8
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.2.0
imblearn  # 用于 SMOTE
albumentations  # 数据增强
